Move envronments into a separate file
